Restructure project, add redux dependencies.
