+++ aoi-ui/src/header-content/header-content.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import './header-content.css';
+import '../css/header-content.css';
 
 const headerContent = () => {
     return (<div className="App">
